requirements.txt
django>=5.2.8
djangorestframework>=3.16.1
django-filter==25.2

docker>=7.1.0

djangorestframework-simplejwt>=5.5.1

psycopg[binary]>=3.2.10
dj-database-url==3.0.1

python-dotenv==1.1.1
pydantic==2.12.2
pydantic-settings==2.11.0

reportlab>=4.4.4
openpyxl>=3.1.5
python-docx>=1.2.0

python-dateutil==2.9.0.post0

celery[redis]>=5.5.3
django-celery-beat>=2.8.1
django-celery-results>=2.6.0
